voice-agents/language_switching/pyproject.toml
[project]
name = "atoms-multi-node-pipeline"
version = "0.1.0"
description = "Chained node processing with event pipelines"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "smallestai>=4.2.0",
    "python-dotenv",
    "loguru",
]
